cart controller

+++ src/main/kotlin/com/mc851/xcommerce/clients/ProductClient.kt
@@ -16,7 +16,7 @@ interface ProductClient {
 
     fun search(text: String): List<ProductApi>
 
-    fun reserve(id: UUID, quantity: Int): Boolean
+    fun reserve(id: UUID, quantity: Long): Boolean
 
-    fun release(id: UUID, quantity: Int): Boolean
+    fun release(id: UUID, quantity: Long): Boolean
 }
